+++ examples/cloudknot_example.py
@@ -44,14 +44,14 @@ def afq_process_subject(subject):
     import s3fs
     # all imports must be at the top of the function
     # cloudknot installs the appropriate packages from pip
-    import AFQ.data as afqd
+    import AFQ.s3bids as afs
     from AFQ.api.group import GroupAFQ
     import AFQ.definitions.mask as afm
 
     # Download the given subject to your local machine from s3
     # Can find subjects more easily if they are specified in a
     # BIDS participants.tsv file, even if it is sparse
-    study_ixi = afqd.S3BIDSStudy(
+    study_ixi = afs.S3BIDSStudy(
         "my_study",
         "my_study_bucket",
         "my_study_prefix",
